+++ CHANGELOG.md
@@ -5,6 +5,10 @@ NOWnews-api change logs
 ## [Unreleased]
 
 ## 1.0.55 - 2018-01-18
+### Fixed
+- 修正抓子網的新聞只抓3小時內建立的新聞, 每三分鐘確認一次不變 @wb
+- news 建立 feedUniqKey index @wb
+
 ### Added
 - 分眾頻道匯出 RSS(PINKNOW, IFUNNOW, PETSMAO, PLAYNOW ,SIGHT, SPORTNOW, BOBEE) @wb
 - 分眾頻道匯進主站(PINKNOW, IFUNNOW, PETSMAO, PLAYNOW ,SIGHT, SPORTNOW, BOBEE) @wb